Added View Binding and Updated Gradle
I replaced FindViewById with View Binding and upgraded the application to
the newest dependencies. This is a more efficient way of designing the
software.

MyAdapter and ResultsActivity is commented out though, which means the
system is not stable. Once I fix those problems I will create a new
commit and merge it into dev.