Changed API and Added Search Functionality
I did a lot of work for this commit. Completely removed spoonacular
from the application, got the search bar functioning, and got the
queries to be dynamic for the Edamam API. It complained more than
Spoonacular.. I also adjusted the data classes and added some to
read the data from the JSON data I retrieve.