Implementation of Accounts and Amplify
I've been working on this project regularly. So There have been a lot of
changes. I didn't push because I wanted to not expose anything I didn't
want to from Amplify first. Then I realised Amplify already handled that
in the .gitignore...

Anyway, I added the sign in and sign up activity and layout files. I
also adjusted the colours for the app, removing the purple from the
accents. It looks much better.

The accounts are kind of functional. I set up the datastore on AWS and
was able to write data directly from the application, but now I have an
unknown issue where authentication is not working because I'm not signed
in. I hope to get past this issue today and create users for AWS
Cognito.