Settling on Firebase - Amplify Partially Intact
I initially wanted to work with Amplify, but it proved to not be fit for
a smallscale application like Refreshments. If I am to deploy the
project, I should keep it simple as my database is not too complex.

This commit includes basic integrations of Firebase for users with
security throughout the application as the application is in use.