Streamlining Toolbar
The toolbar needed some work before I got on to the accounts section,
including how I navigated back to home. Now, an arrow back button takes
the user back home. Also, the search bar is much nicer how it's
implemented. I'm utilising OnQueryTextListener to grab the data from the
user. It also is not janky anymore.

I still need to get that data from the search bar to the results, but
that's deviating from the accounts-feature branch. I'll get around to
that when I'm finished with the accounts