A Lot of Additions To Database and Mobile Backend
Account now has a profile picture for each user. This profile picture is
able to be changed and is updated dynamically. The same goes for a first
name, last name, a username, and a bio.

Some other tweaks to the mobile app were made. Mainly making it look
more modern and up to date with current trends. I also introduced
coroutines to many of the network call functions to allow asynchronous
programming. Fits really well with how I'm interacting with the
database.

I need to get better with pushing up changes frequently. I have
been working on this throughout but got stuck a few times. I should try
to keep it stable when I can and push it up anyway