Filters Integrations and Fixes
Large commit consisting of a lot of additions and changes. Quick filters
is a last minute new feature before app release. I thought that manually
putting in features could get a bit much, so this is a shortcut for
vegans, vegetarians, celliacs, etc.

I created a few adapters, one for filters, another for quick filters. To
accomodate this expansion, I renamed my adapter to SearchAdapter as
well. There is a whole new screen and activtiy obviously for the quick
filters.

One thing that really stood out to me was creating a
QuickFiltersViewModel. Since I'm juggling database data that change their
states frequently, I needed data which existed eternally once the search
or quick filters activity was created. This does work well with coroutines.

I tidied up some code as well across the application since the flow
evolved. The functionality is in a state which I'm happy with, next is
making it look pretty!