Dark Mode Support, UI Additions, and API Fix
I added dark mode support to the app. Icons change depending on what is
device mode it is. This is not integrated into the app itself. The UI in
general also looks nicer too. Recycler views are have more content and
show as expected.

There was a bug introduced that I wasn't aware of when designing the
APIInterface. When a user is designing the filters, they may have none
at all, which creates null for the query being sent. Now there are more
functions in the interface to accomodate this issue.

There some more UI changes coming later. This version is currently being
tested by Google Play Developers for review to the Play Store. Feedback
will then shape the design of the app.